Refactor App component and add new features
- Removed unused imports and consolidated utility functions into discoveredCombinations.ts
- Introduced ClearDiscoveredDialog component for clearing discovered combinations with confirmation dialog
- Created CombinationSection component to handle rendering of combinations and their discovered states
- Implemented ElementSearch component for searching and selecting elements
- Added styled components for better UI consistency
- Created tests for ClearDiscoveredDialog, CombinationSection, ElementSearch, InstallPromptBanner, and PrimaryElement
- Added unit tests for discoveredCombinations utility functions
