+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "little-alchemist-2-cheats",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "homepage": "https://manix84.github.io/little-alchemist-2-cheats",
   "private": true,
   "dependencies": {
