+++ src/components/ClearDiscoveredDialog.test.tsx
@@ -1,0 +1,18 @@
+import { fireEvent, render, screen } from "@testing-library/react";
+import { expect, test, vi } from "vitest";
+import { ClearDiscoveredDialog } from "./ClearDiscoveredDialog";
+
+test("confirms or cancels clearing discovered combinations", () => {
+  const clearDiscoveredCombinations = vi.fn();
+  const closeDialog = vi.fn();
+
+  render(<ClearDiscoveredDialog clearDiscoveredCombinations={clearDiscoveredCombinations} closeDialog={closeDialog} />);
+
+  expect(screen.getByRole("dialog", { name: /clear discovered combinations/i })).toBeInTheDocument();
+
+  fireEvent.click(screen.getByRole("button", { name: /cancel/i }));
+  fireEvent.click(screen.getByRole("button", { name: /clear all/i }));
+
+  expect(closeDialog).toHaveBeenCalledOnce();
+  expect(clearDiscoveredCombinations).toHaveBeenCalledOnce();
+});
